fix(automerge): check branch status again before automerge
Catches an edge case where a PR is configured for automerge and a new update comes right before automerge. If branch protection wasn’t configured, Renovate was using a “stale” branch status to think it’s OK to merge. Added an additional check before automerging that should catch it.

Fixes #3099